+++ NFCCard/App/NFCCardApp.swift
@@ -2,7 +2,8 @@ import SwiftUI
 
 @main
 struct NFCCardApp: App {
-    @StateObject private var scanner = NFCScanner()
+    @Environment(\.scenePhase) private var scenePhase
+    @StateObject private var scanner = NFCScanner(driver: CoreNFCReader())
     @StateObject private var library = CardLibraryStore()
 
     var body: some Scene {
@@ -12,6 +13,9 @@ struct NFCCardApp: App {
                 .environmentObject(library)
                 .task {
                     scanner.library = library
+                }
+                .onChange(of: scenePhase) { phase in
+                    if phase == .background { scanner.enteredBackground() }
                 }
         }
     }